+++ packages/asset-buyer/src/types.ts
@@ -4,7 +4,7 @@ import { BigNumber } from '@0x/utils';
 /**
  * makerAssetData: The assetData representing the desired makerAsset.
  * takerAssetData: The assetData representing the desired takerAsset.
- * networkId: The networkId that the desired orders should be for.
+ * chainId: The chainId that the desired orders should be for.
  */
 export interface OrderProviderRequest {
     makerAssetData: string;
@@ -98,12 +98,12 @@ export interface BuyQuoteExecutionOpts {
 }
 
 /**
- * networkId: The ethereum network id. Defaults to 1 (mainnet).
+ * chainId: The ethereum chain id. Defaults to 1 (mainnet).
  * orderRefreshIntervalMs: The interval in ms that getBuyQuoteAsync should trigger an refresh of orders and order states. Defaults to 10000ms (10s).
  * expiryBufferSeconds: The number of seconds to add when calculating whether an order is expired or not. Defaults to 300s (5m).
  */
 export interface AssetBuyerOpts {
-    networkId: number;
+    chainId: number;
     orderRefreshIntervalMs: number;
     expiryBufferSeconds: number;
 }