+++ packages/contract-wrappers/test/coordinator_wrapper_test.ts
@@ -65,7 +65,7 @@ describe.skip('CoordinatorWrapper', () => {
         const contractAddresses = await migrateOnceAsync();
         await blockchainLifecycle.startAsync();
         const config = {
-            networkId: constants.TESTRPC_NETWORK_ID,
+            chainId: constants.TESTRPC_CHAIN_ID,
             contractAddresses,
             blockPollingIntervalMs: 10,
         };
@@ -111,7 +111,8 @@ describe.skip('CoordinatorWrapper', () => {
         const coordinatorServerConfigs = {
             HTTP_PORT: 3000, // Only used in default instantiation in 0x-coordinator-server/server.js; not used here
             NETWORK_ID_TO_SETTINGS: {
-                [config.networkId]: {
+                // TODO: change to CHAIN_ID_TO_SETTINGS when @0x/coordinator-server is ready
+                [config.chainId]: {
                     FEE_RECIPIENTS: [
                         {
                             ADDRESS: feeRecipientAddressOne,
@@ -138,7 +139,8 @@ describe.skip('CoordinatorWrapper', () => {
                 },
             },
             NETWORK_ID_TO_CONTRACT_ADDRESSES: {
-                [config.networkId]: contractAddresses,
+                // TODO: change to CHAIN_ID_TO_CONTRACT_ADDRESSES when @0x/coordinator-server is ready
+                [config.chainId]: contractAddresses,
             },
             // Optional selective delay on fill requests
             SELECTIVE_DELAY_MS: 0,
@@ -146,7 +148,7 @@ describe.skip('CoordinatorWrapper', () => {
         };
         coordinatorServerApp = await getAppAsync(
             {
-                [config.networkId]: provider,
+                [config.chainId]: provider,
             },
             coordinatorServerConfigs,
             {
@@ -166,7 +168,7 @@ describe.skip('CoordinatorWrapper', () => {
 
         anotherCoordinatorServerApp = await getAppAsync(
             {
-                [config.networkId]: provider,
+                [config.chainId]: provider,
             },
             coordinatorServerConfigs,
             {
@@ -454,7 +456,7 @@ describe.skip('CoordinatorWrapper', () => {
             nock(`${coordinatorEndpoint}${coordinatorPort}`)
                 .post('/v1/request_transaction', () => true)
                 .query({
-                    networkId: 50,
+                    chainId: 1337,
                 })
                 .reply(400, serverValidationError);
         });
@@ -503,7 +505,7 @@ describe.skip('CoordinatorWrapper', () => {
             nock(`${coordinatorEndpoint}${anotherCoordinatorPort}`)
                 .post('/v1/request_transaction', () => true)
                 .query({
-                    networkId: 50,
+                    chainId: 1337,
                 })
                 .reply(200, serverCancellationSuccess);
 
@@ -531,7 +533,7 @@ describe.skip('CoordinatorWrapper', () => {
             nock(`${coordinatorEndpoint}${anotherCoordinatorPort}`)
                 .post('/v1/request_transaction', () => true)
                 .query({
-                    networkId: 50,
+                    chainId: 1337,
                 })
                 .reply(400, serverValidationError);
 
@@ -610,7 +612,7 @@ describe.skip('CoordinatorWrapper', () => {
             nock(`${coordinatorEndpoint}${anotherCoordinatorPort}`)
                 .post('/v1/request_transaction', () => true)
                 .query({
-                    networkId: 50,
+                    chainId: 1337,
                 })
                 .reply(200, serverApprovalSuccess);
 
@@ -639,7 +641,7 @@ describe.skip('CoordinatorWrapper', () => {
             nock(`${coordinatorEndpoint}${anotherCoordinatorPort}`)
                 .post('/v1/request_transaction', () => true)
                 .query({
-                    networkId: 50,
+                    chainId: 1337,
                 })
                 .reply(400, serverValidationError);