+++ packages/contract-wrappers/test/utils/constants.ts
@@ -2,9 +2,8 @@ import { BigNumber } from '@0x/utils';
 
 export const constants = {
     NULL_ADDRESS: '0x0000000000000000000000000000000000000000',
-    ROPSTEN_NETWORK_ID: 3,
-    KOVAN_NETWORK_ID: 42,
-    TESTRPC_NETWORK_ID: 50,
+    ROPSTEN_CHAIN_ID: 3,
+    KOVAN_CHAIN_ID: 42,
     AWAIT_TRANSACTION_MINED_MS: 0,
     KOVAN_RPC_URL: 'https://kovan.infura.io/',
     ROPSTEN_RPC_URL: 'https://ropsten.infura.io/',