+++ packages/migrations/src/utils/token_info.ts
@@ -2,7 +2,7 @@ import { BigNumber, NULL_BYTES } from '@0x/utils';
 
 import { ERC20Token, ERC721Token } from '../types';
 
-export const etherTokenByNetwork: { [networkId: number]: { address: string } } = {
+export const etherTokenByChain: { [chainId: number]: { address: string } } = {
     3: {
         address: '0xc778417e063141139fce010982780140aa0cd5ab',
     },
@@ -12,7 +12,7 @@ export const etherTokenByNetwork: { [networkId: number]: { address: string } } =
     42: {
         address: '0xd0a1e359811322d97991e03f863a0c30c2cf029c',
     },
-    50: {
+    1337: {
         address: '',
     },
 };