+++ packages/testnet-faucets/src/ts/handler.ts
@@ -9,7 +9,7 @@ import {
     SignedOrder,
     Web3ProviderEngine,
 } from '0x.js';
-import { getContractAddressesForNetworkOrThrow } from '@0x/contract-addresses';
+import { getContractAddressesForChainOrThrow } from '@0x/contract-addresses';
 import { NonceTrackerSubprovider, PrivateKeyWalletSubprovider } from '@0x/subproviders';
 import { logUtils } from '@0x/utils';
 import { SupportedProvider, Web3Wrapper } from '@0x/web3-wrapper';
@@ -21,17 +21,17 @@ import { constants } from './constants';
 import { DispatchQueue } from './dispatch_queue';
 import { dispenseAssetTasks } from './dispense_asset_tasks';
 import { rpcUrls } from './rpc_urls';
-import { TOKENS_BY_NETWORK } from './tokens';
+import { TOKENS_BY_CHAIN } from './tokens';
 
-interface NetworkConfig {
+interface ChainConfig {
     dispatchQueue: DispatchQueue;
     web3Wrapper: Web3Wrapper;
     provider: SupportedProvider;
-    networkId: number;
+    chainId: number;
 }
 
-interface ItemByNetworkId<T> {
-    [networkId: string]: T;
+interface ItemByChainId<T> {
+    [chainId: string]: T;
 }
 
 enum RequestedAssetType {
@@ -46,7 +46,7 @@ const ZERO = new BigNumber(0);
 const ASSET_AMOUNT = new BigNumber(0.1);
 
 export class Handler {
-    private readonly _networkConfigByNetworkId: ItemByNetworkId<NetworkConfig> = {};
+    private readonly _chainConfigByChainId: ItemByChainId<ChainConfig> = {};
     private static _createProviderEngine(rpcUrl: string): Web3ProviderEngine {
         if (configs.DISPENSER_PRIVATE_KEY === undefined) {
             throw new Error('Dispenser Private key not found');
@@ -59,24 +59,24 @@ export class Handler {
         return engine;
     }
     constructor() {
-        _.forIn(rpcUrls, (rpcUrl: string, networkIdString: string) => {
+        _.forIn(rpcUrls, (rpcUrl: string, chainIdString: string) => {
             const providerObj = Handler._createProviderEngine(rpcUrl);
             const web3Wrapper = new Web3Wrapper(providerObj);
             // tslint:disable-next-line:custom-no-magic-numbers
-            const networkId = parseInt(networkIdString, 10);
+            const chainId = parseInt(chainIdString, 10);
             const dispatchQueue = new DispatchQueue();
-            this._networkConfigByNetworkId[networkId] = {
+            this._chainConfigByChainId[chainId] = {
                 dispatchQueue,
                 web3Wrapper,
                 provider: providerObj,
-                networkId,
+                chainId,
             };
         });
     }
     public getQueueInfo(_req: express.Request, res: express.Response): void {
         res.setHeader('Content-Type', 'application/json');
-        const queueInfo = _.mapValues(rpcUrls, (_rpcUrl: string, networkId: string) => {
-            const dispatchQueue = this._networkConfigByNetworkId[networkId].dispatchQueue;
+        const queueInfo = _.mapValues(rpcUrls, (_rpcUrl: string, chainId: string) => {
+            const dispatchQueue = this._chainConfigByChainId[chainId].dispatchQueue;
             return {
                 full: dispatchQueue.isFull(),
                 size: dispatchQueue.size(),
@@ -102,36 +102,36 @@ export class Handler {
         await this._dispenseOrderAsync(req, res, RequestedAssetType.ZRX);
     }
     private _dispenseAsset(req: express.Request, res: express.Response, requestedAssetType: RequestedAssetType): void {
-        const networkId = req.params.networkId;
+        const chainId = req.params.chainId;
         const recipient = req.params.recipient;
-        const networkConfig = _.get(this._networkConfigByNetworkId, networkId);
-        if (networkConfig === undefined) {
-            res.status(constants.BAD_REQUEST_STATUS).send('UNSUPPORTED_NETWORK_ID');
+        const chainConfig = _.get(this._chainConfigByChainId, chainId);
+        if (chainConfig === undefined) {
+            res.status(constants.BAD_REQUEST_STATUS).send('UNSUPPORTED_CHAIN_ID');
             return;
         }
         let dispenserTask;
         switch (requestedAssetType) {
             case RequestedAssetType.ETH:
-                dispenserTask = dispenseAssetTasks.dispenseEtherTask(recipient, networkConfig.web3Wrapper);
+                dispenserTask = dispenseAssetTasks.dispenseEtherTask(recipient, chainConfig.web3Wrapper);
                 break;
             case RequestedAssetType.WETH:
             case RequestedAssetType.ZRX:
                 dispenserTask = dispenseAssetTasks.dispenseTokenTask(
                     recipient,
                     requestedAssetType,
-                    networkConfig.networkId,
-                    networkConfig.provider,
+                    chainConfig.chainId,
+                    chainConfig.provider,
                 );
                 break;
             default:
                 throw new Error(`Unsupported asset type: ${requestedAssetType}`);
         }
-        const didAddToQueue = networkConfig.dispatchQueue.add(dispenserTask);
+        const didAddToQueue = chainConfig.dispatchQueue.add(dispenserTask);
         if (!didAddToQueue) {
             res.status(constants.SERVICE_UNAVAILABLE_STATUS).send('QUEUE_IS_FULL');
             return;
         }
-        logUtils.log(`Added ${recipient} to queue: ${requestedAssetType} networkId: ${networkId}`);
+        logUtils.log(`Added ${recipient} to queue: ${requestedAssetType} chainId: ${chainId}`);
         res.status(constants.SUCCESS_STATUS).end();
     }
     private async _dispenseOrderAsync(
@@ -139,19 +139,19 @@ export class Handler {
         res: express.Response,
         requestedAssetType: RequestedAssetType,
     ): Promise<void> {
-        const networkConfig = _.get(this._networkConfigByNetworkId, req.params.networkId);
-        if (networkConfig === undefined) {
-            res.status(constants.BAD_REQUEST_STATUS).send('UNSUPPORTED_NETWORK_ID');
+        const chainConfig = _.get(this._chainConfigByChainId, req.params.chainId);
+        if (chainConfig === undefined) {
+            res.status(constants.BAD_REQUEST_STATUS).send('UNSUPPORTED_CHAIN_ID');
             return;
         }
         res.setHeader('Content-Type', 'application/json');
-        const makerTokenIfExists = _.get(TOKENS_BY_NETWORK, [networkConfig.networkId, requestedAssetType]);
+        const makerTokenIfExists = _.get(TOKENS_BY_CHAIN, [chainConfig.chainId, requestedAssetType]);
         if (makerTokenIfExists === undefined) {
             throw new Error(`Unsupported asset type: ${requestedAssetType}`);
         }
         const takerTokenSymbol =
             requestedAssetType === RequestedAssetType.WETH ? RequestedAssetType.ZRX : RequestedAssetType.WETH;
-        const takerTokenIfExists = _.get(TOKENS_BY_NETWORK, [networkConfig.networkId, takerTokenSymbol]);
+        const takerTokenIfExists = _.get(TOKENS_BY_CHAIN, [chainConfig.chainId, takerTokenSymbol]);
         if (takerTokenIfExists === undefined) {
             throw new Error(`Unsupported asset type: ${takerTokenSymbol}`);
         }
@@ -160,7 +160,7 @@ export class Handler {
         const takerAssetAmount = Web3Wrapper.toBaseUnitAmount(ASSET_AMOUNT, takerTokenIfExists.decimals);
         const makerAssetData = assetDataUtils.encodeERC20AssetData(makerTokenIfExists.address);
         const takerAssetData = assetDataUtils.encodeERC20AssetData(takerTokenIfExists.address);
-        const contractAddresses = getContractAddressesForNetworkOrThrow(networkConfig.networkId);
+        const contractAddresses = getContractAddressesForChainOrThrow(chainConfig.chainId);
         const order: Order = {
             makerAddress: configs.DISPENSER_ADDRESS,
             takerAddress: req.params.recipient as string,
@@ -180,11 +180,11 @@ export class Handler {
                 .div(1000)
                 .integerValue(BigNumber.ROUND_FLOOR),
             exchangeAddress: contractAddresses.exchange,
-            chainId: networkConfig.networkId,
+            chainId: chainConfig.chainId,
         };
         const orderHash = orderHashUtils.getOrderHashHex(order);
         const signature = await signatureUtils.ecSignHashAsync(
-            networkConfig.web3Wrapper.getProvider(),
+            chainConfig.web3Wrapper.getProvider(),
             orderHash,
             configs.DISPENSER_ADDRESS,
         );