+++ python-packages/contract_wrappers/test/test_base_contract_method.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from zero_ex.contract_addresses import network_to_addresses, NetworkId
+from zero_ex.contract_addresses import chain_to_addresses, ChainId
 from zero_ex.contract_wrappers.bases import ContractMethod
 
 
@@ -11,5 +11,5 @@ def contract_wrapper(ganache_provider):
     """Get a ContractMethod instance for testing."""
     return ContractMethod(
         web3_or_provider=ganache_provider,
-        contract_address=network_to_addresses(NetworkId.GANACHE).ether_token,
+        contract_address=chain_to_addresses(ChainId.GANACHE).ether_token,
     )